+++ sotoki/static/main.js
@@ -1,5 +1,0 @@
-$('span.date').each(function(index, elt) {
-    var element = $(elt);
-    element.html(moment(element.html()).fromNow());
-});
-